update
I worked for as long as I could tonight.  This code has yet to be compiled because i was unable to figure out how to get c++ 14 up and running on the library PC computers.  I believe that most of the code will work.  Tomorrow I should have more free time to continue working on the project.